\name{muscleResult}
\alias{muscleResult}
\docType{data}
\title{ data set contains muscle result}
\description{
 Results from running the muscle example in \code{\link{glmmBUGS}}.
}
\usage{data(muscleResult)}
\format{A list as returned by the \code{\link[R2WinBUGS:bugs]{bugs}} function.
}
\details{
See \code{\link{glmmBUGS}} and \code{\link[nlme:Muscle]{Muscle}}

}
 
\examples{
data(muscleResult)

}
\keyword{datasets}
