\name{restoreParams}
\alias{restoreParams}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Reparametrise bugs output}
\description{
  Undoes the parametrisation used in \code{\link{writeBugsModel}}, and gives the original names to random effect levels.
}
\usage{
restoreParams(bugsResult, ragged = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{bugsResult}{Output from \code{\link[R2WinBUGS:bugs]{bugs}}, using a ragged array generated by \code{\link{winBugsRaggedArray}} and a model generated by \code{\link{writeBugsModel}}}
  \item{ragged}{The ragged array used to call \code{\link[R2WinBUGS:bugs]{bugs}}}
}
\details{
 }
\value{
  A list where each element is a matrix or an array.  
  The first dimension is the number of realisations, 
  the second the number of chains, and for vector-valued parameters and 
  random effects, the third dimension is the length of the parameter.
  
  
If the model contains a spatial component, 
the result will have list entries the following:

\item{ReffectSpatial}{The spatial random effect for each region}
\item{Reffect}{The log relative risk, which is the sum of the spatial and non-spatial random effect}, 
\item{FittedRateeffect}{The incidence rate for each region (if the offset in the model is total population), 
or the incident rate relative to the population average (if the offset is an expected count).  Note that 
these rates are on the natural scale rather than the log scale.}
}
\references{}
\author{Patrick Brown \code{patrick.brown@utoronto.ca}}
\note{}
\seealso{\code{\link[R2WinBUGS:bugs]{bugs}}}
\examples{
                        
}
