\name{getRaggedSeq}
\alias{getRaggedSeq}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Get one sequence for a ragged array}
\description{
This function is called by \code{\link{winBugsRaggedArray}}
}
\usage{
getRaggedSeq(data)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data}{a data frame with two columns}
}
\details{
  
}
\value{
 The ragged sequence
}
\references{}
\author{Patrick Brown \email{patrick.brown@utoronto.ca}}
\note{
}
\seealso{\code{\link{winBugsRaggedArray}}}
\examples{
}