\name{restoreParams}
\alias{restoreParams}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Reparametrise bugs output}
\description{
  Undoes the parametrisation used in \code{\link{writeBugsModel}}, and gives the original names to random effect levels.
}
\usage{
restoreParams(bugsResult, ragged = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{bugsResult}{Output from \code{\link[R2WinBUGS:bugs]{bugs}}, using a ragged array generated by \code{\link{winBugsRaggedArray}} and a model generated by \code{\link{writeBugsModel}}}
  \item{ragged}{The ragged array used to call \code{\link[R2WinBUGS:bugs]{bugs}}}
}
\details{
}
\value{
  A list where each element is a matrix or an array.  The first dimension is the number of realisations, the second the number of chains, and for vector-valued parameters and random effects, the third dimension is the length of the parameter..
}
\references{}
\author{Patrick Brown \code{patrick.brown@utoronto.ca}}
\note{}
\seealso{\code{\link[R2WinBUGS:bugs]{bugs}}}
\examples{
                        
}
